Change total_score and score columns to decimal(8,2) for PostgreSQL compatibility

+++ database/migrations/2025_10_11_155717_create_exam_results_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('student_id')->constrained('users')->onDelete('cascade');
             $table->foreignId('exam_id')->constrained()->onDelete('cascade');
-            $table->integer('total_score')->default(0); // tổng điểm
+            $table->decimal('total_score', 8, 2)->default(0); // tổng điểm
             $table->integer('correct_count')->default(0); // số câu đúng
             $table->integer('total_questions')->default(0); // tổng số câu
             $table->timestamp('submitted_at')->nullable(); // thời gian nộp bài
